+++ server/src/main/java/com/linbit/linstor/netcom/PeerTask.java
@@ -14,6 +14,7 @@ import com.linbit.linstor.security.AccessDeniedException;
 import com.linbit.linstor.security.Privilege;
 import com.linbit.linstor.utils.externaltools.ExtToolsManager;
 
+import javax.annotation.Nullable;
 import javax.net.ssl.SSLException;
 
 import java.io.ByteArrayInputStream;
@@ -53,6 +54,13 @@ public class PeerTask implements Peer
     public String getId()
     {
         return peerId;
+    }
+
+    @Nullable
+    @Override
+    public InetSocketAddress getHostAddr()
+    {
+        return null;
     }
 
     @Override